TMS-240: Refactor (non-browser) tests so we can have several test modules.
