+++ src/zeit/find/browser/README.txt
@@ -166,7 +166,7 @@ Search
 The search view returns all data for rendering the result:
 
 >>> import zeit.find.tests
->>> zeit.find.tests.LAYER.set_result('zeit.find', 'testdata/obama.json')
+>>> zeit.find.tests.LAYER.set_result('zeit.find', 'tests/data/obama.json')
 >>> browser.open('/++skin++cms/search_result?fulltext=Obama')
 >>> result = json.loads(browser.contents)
 >>> pprint.pprint(result)
